bridge: divide token and bridge modules
